+++ HISTORY.md
@@ -5,5 +5,5 @@
 ## 2.0.0
 
 - refactor code
-- Add `prefixCls` props
+- Add `prefixCls` `className` `style` props.
 - Add `strokeLinecap` for shape of end in progress bar.